+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   annotations are on the page
 * [HexaPDF::Type::AcroForm::ButtonField#create_appearances] to avoid creating
   appearances in case of as-yet unresolved references to existing appearances
+* `HexaPDF::Tokenizer#parse_number` to treat invalid indirect object references
+  with an object number of 0 as null values
 
 
 ## 0.27.0 - 2022-11-18